browser-use-automation/requirements.txt
browser-use-sdk>=2.0.0
python-dotenv>=1.0.1
pydantic>=2.0.0
